@@ -9,11 +9,9 @@
     "@testing-library/user-event": "^13.2.1",
     "@types/jest": "^27.0.1",
     "@types/node": "^16.7.13",
-    "@types/react": "^17.0.20",
-    "@types/react-dom": "^17.0.9",
-    "react": "^17.0.2",
-    "react-animate-on-scroll": "^2.1.5",
-    "react-dom": "^17.0.2",
+    "aos": "^2.3.4",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-google-recaptcha": "^2.1.0",
     "react-router-dom": "^6.2.1",
     "react-router-hash-link": "^2.4.3",
@@ -50,7 +48,10 @@
     ]
   },
   "devDependencies": {
+    "@types/aos": "^3.0.7",
+    "@types/react": "^18.3.23",
     "@types/react-animate-on-scroll": "^2.1.5",
+    "@types/react-dom": "^18.3.7",
     "@types/react-google-recaptcha": "^2.1.2",
     "@types/react-router-hash-link": "^2.4.4",
     "@types/styled-components": "^5.1.18",
