@@ -1,22 +1,20 @@
-import { Container, ContainerSucces } from './styles'
-import { useForm, ValidationError } from '@formspree/react'
-import { toast, ToastContainer } from 'react-toastify'
-import ReCAPTCHA from 'react-google-recaptcha'
-import { useEffect, useState } from 'react'
-import validator from 'validator'
+import { Container, ContainerSucces } from './styles';
+import { useForm, ValidationError } from '@formspree/react';
+import { toast, ToastContainer } from 'react-toastify';
+import ReCAPTCHA from 'react-google-recaptcha';
+import { useEffect, useState } from 'react';
+import validator from 'validator';
 
 export function Form() {
-  const [state, handleSubmit] = useForm('xknkpqry')
-  const [validEmail, setValidEmail] = useState(false)
-  const [isHuman, setIsHuman] = useState(false)
-  const [message, setMessage] = useState('')
+  const [state, handleSubmit] = useForm('xknkpqry');
+  const [validEmail, setValidEmail] = useState(false);
+  const [isHuman, setIsHuman] = useState(false);
+  const [message, setMessage] = useState('');
+
   function verifyEmail(email: string) {
-    if (validator.isEmail(email)) {
-      setValidEmail(true)
-    } else {
-      setValidEmail(false)
-    }
+    setValidEmail(validator.isEmail(email));
   }
+
   useEffect(() => {
     if (state.succeeded) {
       toast.success('Email successfully sent!', {
@@ -25,24 +23,26 @@ export function Form() {
         closeOnClick: true,
         hideProgressBar: false,
         toastId: 'succeeded',
-      })
+      });
     }
-  })
+  }, [state.succeeded]);
+
   if (state.succeeded) {
     return (
       <ContainerSucces>
         <h3>Thanks for getting in touch!</h3>
         <button
           onClick={() => {
-            window.scrollTo({ top: 0, behavior: 'smooth' })
+            window.scrollTo({ top: 0, behavior: 'smooth' });
           }}
         >
           Back to the top
         </button>
         <ToastContainer />
       </ContainerSucces>
-    )
+    );
   }
+
   return (
     <Container>
       <h2>Get in touch using the form</h2>
@@ -52,32 +52,26 @@ export function Form() {
           id="email"
           type="email"
           name="email"
-          onChange={(e) => {
-            verifyEmail(e.target.value)
-          }}
+          onChange={(e) => verifyEmail(e.target.value)}
           required
         />
         <ValidationError prefix="Email" field="email" errors={state.errors} />
+
         <textarea
           required
           placeholder="Send a message to get started."
           id="message"
           name="message"
-          onChange={(e) => {
-            setMessage(e.target.value)
-          }}
+          value={message}
+          onChange={(e) => setMessage(e.target.value)}
         />
-        <ValidationError
-          prefix="Message"
-          field="message"
-          errors={state.errors}
-        />
+        <ValidationError prefix="Message" field="message" errors={state.errors} />
+
         <ReCAPTCHA
           sitekey="6Lfj9NYfAAAAAP8wPLtzrsSZeACIcGgwuEIRvbSg"
-          onChange={(e) => {
-            setIsHuman(true)
-          }}
-        ></ReCAPTCHA>
+          onChange={() => setIsHuman(true)}
+        />
+
         <button
           type="submit"
           disabled={state.submitting || !validEmail || !message || !isHuman}
@@ -87,5 +81,5 @@ export function Form() {
       </form>
       <ToastContainer />
     </Container>
-  )
+  );
 }
